fdf changelog (1) fdf renders the map in correct order when rotational transform is applied. (2) scale up & down towards center of the display, not the center of the map. (3) file seperation (4) norminette 3.3.40 checked. (5) frametime function output  information on the display.

+++ core2/fdf/src/fdf_manage_map.c
@@ -6,7 +6,7 @@
 /*   By: hseong <hseong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/27 18:31:12 by hseong            #+#    #+#             */
-/*   Updated: 2022/03/28 20:40:34 by hseong           ###   ########.fr       */
+/*   Updated: 2022/03/29 18:30:40 by hseong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@ void	fdf_align_map(t_map_info *map_info)
 	map_info->hor_angle = HOR_ANGLE;
 	map_info->ver_angle = VER_ANGLE;
 	map_info->hor_scale = (double)SCREEN_WIDTH
-			/ (map_info->col + map_info->row);
+		/ (map_info->col + map_info->row);
 	map_info->ver_scale = (double)SCREEN_HEIGHT
-			/ (double)(map_info->col + map_info->row + map_info->max_height / 2);
+		/ (map_info->col + map_info->row + map_info->max_height / 2.0);
 	map_info->x0 = SCREEN_WIDTH / 2;
 	map_info->y0 = SCREEN_HEIGHT * 10 / 19;
 }
